Create django_mongodb_backend.functions package
