--- django_mongodb_backend/functions.py
@@ -1,505 +1,0 @@
-from datetime import datetime
-from functools import partialmethod
-
-from django.conf import settings
-from django.db import NotSupportedError
-from django.db.models import DateField, DateTimeField, TimeField
-from django.db.models.expressions import Func
-from django.db.models.functions import JSONArray
-from django.db.models.functions.comparison import Cast, Coalesce, Greatest, Least, NullIf
-from django.db.models.functions.datetime import (
-    Extract,
-    ExtractDay,
-    ExtractHour,
-    ExtractIsoWeekDay,
-    ExtractIsoYear,
-    ExtractMinute,
-    ExtractMonth,
-    ExtractQuarter,
-    ExtractSecond,
-    ExtractWeek,
-    ExtractWeekDay,
-    ExtractYear,
-    Now,
-    TruncBase,
-    TruncDate,
-    TruncTime,
-)
-from django.db.models.functions.math import (
-    Ceil,
-    Cot,
-    Degrees,
-    Log,
-    Power,
-    Radians,
-    Random,
-    Round,
-    Sign,
-)
-from django.db.models.functions.text import (
-    MD5,
-    SHA256,
-    Concat,
-    ConcatPair,
-    Left,
-    Length,
-    Lower,
-    LPad,
-    LTrim,
-    Repeat,
-    Replace,
-    Reverse,
-    Right,
-    RPad,
-    RTrim,
-    StrIndex,
-    Substr,
-    Trim,
-    Upper,
-)
-
-from .query_utils import process_lhs
-
-MONGO_OPERATORS = {
-    Ceil: "ceil",
-    Coalesce: "ifNull",
-    Degrees: "radiansToDegrees",
-    Greatest: "max",
-    Least: "min",
-    Power: "pow",
-    Radians: "degreesToRadians",
-    Random: "rand",
-}
-EXTRACT_OPERATORS = {
-    ExtractDay.lookup_name: "dayOfMonth",
-    ExtractHour.lookup_name: "hour",
-    ExtractIsoWeekDay.lookup_name: "isoDayOfWeek",
-    ExtractIsoYear.lookup_name: "isoWeekYear",
-    ExtractMinute.lookup_name: "minute",
-    ExtractMonth.lookup_name: "month",
-    ExtractSecond.lookup_name: "second",
-    ExtractWeek.lookup_name: "isoWeek",
-    ExtractWeekDay.lookup_name: "dayOfWeek",
-    ExtractYear.lookup_name: "year",
-}
-
-
-def cast(self, compiler, connection):
-    output_type = connection.data_types[self.output_field.get_internal_type()]
-    lhs_mql = process_lhs(self, compiler, connection, as_expr=True)[0]
-    if max_length := self.output_field.max_length:
-        lhs_mql = {"$substrCP": [lhs_mql, 0, max_length]}
-    # Skip the conversion for "object" as it doesn't need to be transformed for
-    # interpretation by JSONField, which can handle types including int,
-    # object, or array.
-    if output_type != "object":
-        lhs_mql = {"$convert": {"input": lhs_mql, "to": output_type}}
-    if decimal_places := getattr(self.output_field, "decimal_places", None):
-        lhs_mql = {"$trunc": [lhs_mql, decimal_places]}
-    return lhs_mql
-
-
-def concat(self, compiler, connection):
-    return self.get_source_expressions()[0].as_mql(compiler, connection, as_expr=True)
-
-
-def concat_pair(self, compiler, connection):
-    # null on either side results in null for expression, wrap with coalesce.
-    coalesced = self.coalesce()
-    return super(ConcatPair, coalesced).as_mql_expr(compiler, connection)
-
-
-def cot(self, compiler, connection):
-    lhs_mql = process_lhs(self, compiler, connection, as_expr=True)
-    return {"$divide": [1, {"$tan": lhs_mql}]}
-
-
-def _get_extract_timezone(self):
-    tzname = self.get_tzname()
-    # Django formats fixed offset zones as "UTC+HH:MM" but MongoDB only accepts
-    # the bare offset form "+HH:MM" / "-HH:MM".
-    if tzname and tzname.startswith(("UTC+", "UTC-")):
-        return tzname[3:]
-    return tzname
-
-
-def extract(self, compiler, connection):
-    lhs_mql = process_lhs(self, compiler, connection, as_expr=True)
-    # ExtractQuarter lacks a built-in operator.
-    if self.lookup_name == "quarter":
-        return extract_quarter(self, compiler, connection)
-    operator = EXTRACT_OPERATORS.get(self.lookup_name)
-    if operator is None:
-        raise NotSupportedError(f"{self.__class__.__name__} is not supported.")
-    if timezone := _get_extract_timezone(self):
-        lhs_mql = {"date": lhs_mql, "timezone": timezone}
-    return {f"${operator}": lhs_mql}
-
-
-def extract_quarter(self, compiler, connection):
-    lhs_mql = process_lhs(self, compiler, connection, as_expr=True)
-    if timezone := _get_extract_timezone(self):
-        lhs_mql = {"date": lhs_mql, "timezone": timezone}
-    return {"$ceil": {"$divide": [{"$month": lhs_mql}, 3]}}
-
-
-def func(self, compiler, connection):
-    lhs_mql = process_lhs(self, compiler, connection, as_expr=True)
-    if self.function is None:
-        raise NotSupportedError(f"{self.__class__.__name__} may need an as_mql() method.")
-    operator = MONGO_OPERATORS.get(self.__class__, self.function.lower())
-    return {f"${operator}": lhs_mql}
-
-
-def hash_func(algorithm):
-    def wrapped(self, compiler, connection):
-        if not connection.features.is_mongodb_8_3:
-            raise NotSupportedError(f"{self.__class__.__name__} requires MongoDB 8.3+.")
-        lhs_mql = process_lhs(self, compiler, connection, as_expr=True)
-        return {
-            "$cond": {
-                "if": {"$eq": [lhs_mql, None]},
-                "then": None,  # Return null for null input.
-                "else": {
-                    "$toLower": {
-                        "$convert": {
-                            "input": {"$hash": {"input": lhs_mql, "algorithm": algorithm}},
-                            "to": "string",
-                            "format": "hex",
-                        }
-                    }
-                },
-            }
-        }
-
-    return wrapped
-
-
-def left(self, compiler, connection):
-    return self.get_substr().as_mql(compiler, connection, as_expr=True)
-
-
-def right(self, compiler, connection):
-    expression, length = (
-        expr.as_mql(compiler, connection, as_expr=True) for expr in self.get_source_expressions()
-    )
-    # Calculate substring's start value by subtracting the requested length
-    # from the total length of the string.
-    start = {"$max": [{"$subtract": [{"$strLenCP": expression}, length]}, 0]}
-    return {
-        "$cond": {
-            "if": {"$eq": [length, None]},
-            # $substrCP returns "" for null length; return null instead.
-            "then": None,
-            "else": {"$substrCP": [expression, start, length]},
-        }
-    }
-
-
-def length(self, compiler, connection):
-    # Check for null first since $strLenCP only accepts strings.
-    lhs_mql = process_lhs(self, compiler, connection, as_expr=True)
-    return {"$cond": {"if": {"$eq": [lhs_mql, None]}, "then": None, "else": {"$strLenCP": lhs_mql}}}
-
-
-def log(self, compiler, connection):
-    # This function is usually log(base, num) but on MongoDB it's log(num, base).
-    clone = self.copy()
-    clone.set_source_expressions(self.get_source_expressions()[::-1])
-    return func(clone, compiler, connection)
-
-
-def now(self, compiler, connection):  # noqa: ARG001
-    return "$$NOW"
-
-
-def null_if(self, compiler, connection):
-    """Return None if expr1==expr2 else expr1."""
-    expr1, expr2 = (
-        expr.as_mql(compiler, connection, as_expr=True) for expr in self.get_source_expressions()
-    )
-    return {"$cond": {"if": {"$eq": [expr1, expr2]}, "then": None, "else": expr1}}
-
-
-def pad(is_left=False):
-    def wrapped(self, compiler, connection):
-        expression, length, fill_text = process_lhs(self, compiler, connection, as_expr=True)
-        # Use $ifNull to avoid $strLenCP failing on null during pipeline
-        # optimization.
-        str_len = {"$strLenCP": {"$ifNull": [expression, ""]}}
-        pad_needed = {"$subtract": [length, str_len]}
-        # Repeat fill_text enough times to cover the padding, then trim to
-        # pad_needed characters using $substrCP.
-        repeated_fill = {
-            "$reduce": {
-                "input": {
-                    "$range": [
-                        0,
-                        {
-                            "$toInt": {
-                                "$ceil": {
-                                    "$divide": [
-                                        pad_needed,
-                                        {"$max": [{"$strLenCP": {"$ifNull": [fill_text, " "]}}, 1]},
-                                    ]
-                                }
-                            }
-                        },
-                    ]
-                },
-                "initialValue": "",
-                "in": {"$concat": ["$$value", fill_text]},
-            }
-        }
-        padding = {"$substrCP": [repeated_fill, 0, pad_needed]}
-        padded = {"$concat": [padding, expression] if is_left else [expression, padding]}
-        return {
-            "$cond": {
-                "if": {"$or": [{"$eq": [expression, None]}, {"$eq": [length, None]}]},
-                "then": None,  # Return null for null inputs.
-                "else": {
-                    "$cond": {
-                        "if": {"$gte": [str_len, length]},
-                        "then": {"$substrCP": [expression, 0, length]},
-                        "else": padded,
-                    }
-                },
-            }
-        }
-
-    return wrapped
-
-
-def preserve_null(operator):
-    # If the argument is null, the function should return null, not
-    # $toLower/Upper's behavior of returning an empty string.
-    def wrapped(self, compiler, connection):
-        lhs_mql = process_lhs(self, compiler, connection, as_expr=True)
-        return {
-            "$cond": {
-                "if": connection.mongo_expr_operators["isnull"](lhs_mql, True),
-                "then": None,
-                "else": {f"${operator}": lhs_mql},
-            }
-        }
-
-    return wrapped
-
-
-def repeat(self, compiler, connection):
-    expression, number = process_lhs(self, compiler, connection, as_expr=True)
-    return {
-        "$cond": {
-            "if": {"$or": [{"$eq": [expression, None]}, {"$eq": [number, None]}]},
-            "then": None,  # Return null if any inputs are null.
-            "else": {
-                "$reduce": {
-                    # $ifNull needed because $range fails during MongoDB
-                    # pipeline optimization when number is null, even inside a
-                    # $cond "else" branch.
-                    "input": {"$range": [0, {"$ifNull": [number, 0]}]},
-                    "initialValue": "",
-                    "in": {"$concat": ["$$value", expression]},
-                }
-            },
-        }
-    }
-
-
-def replace(self, compiler, connection):
-    expression, text, replacement = process_lhs(self, compiler, connection, as_expr=True)
-    return {"$replaceAll": {"input": expression, "find": text, "replacement": replacement}}
-
-
-def reverse(self, compiler, connection):
-    lhs_mql = process_lhs(self, compiler, connection, as_expr=True)
-    chars = {
-        "$map": {
-            "input": {"$range": [0, {"$strLenCP": lhs_mql}]},
-            "as": "i",
-            "in": {"$substrCP": [lhs_mql, "$$i", 1]},
-        }
-    }
-    return {
-        "$cond": {
-            "if": {"$eq": [lhs_mql, None]},
-            "then": None,  # Return null for null input.
-            "else": {
-                "$reduce": {
-                    "input": {"$reverseArray": chars},
-                    "initialValue": "",
-                    "in": {"$concat": ["$$value", "$$this"]},
-                }
-            },
-        }
-    }
-
-
-def round_(self, compiler, connection):
-    # Round needs its own function because it's a special case that inherits
-    # from Transform but has two arguments.
-    return {
-        "$round": [
-            expr.as_mql(compiler, connection, as_expr=True)
-            for expr in self.get_source_expressions()
-        ]
-    }
-
-
-def sign(self, compiler, connection):
-    lhs_mql = process_lhs(self, compiler, connection, as_expr=True)
-    return {
-        "$cond": {
-            "if": {"$eq": [lhs_mql, None]},
-            "then": None,  # Return null for null input.
-            "else": {
-                "$cond": {
-                    "if": {"$eq": [lhs_mql, 0]},
-                    "then": 0,  # Return zero for zero input.
-                    "else": {"$cmp": [lhs_mql, 0]},  # Otherwise, +1 or -1.
-                },
-            },
-        }
-    }
-
-
-def str_index(self, compiler, connection):
-    lhs = process_lhs(self, compiler, connection, as_expr=True)
-    # StrIndex should be 0-indexed (not found) but it's -1-indexed on MongoDB.
-    return {"$add": [{"$indexOfCP": lhs}, 1]}
-
-
-def substr(self, compiler, connection):
-    lhs = process_lhs(self, compiler, connection, as_expr=True)
-    # The starting index is zero-indexed on MongoDB rather than one-indexed.
-    lhs[1] = {"$add": [lhs[1], -1]}
-    # If no limit is specified, use the length of the string since $substrCP
-    # requires one.
-    if len(lhs) == 2:
-        lhs.append({"$strLenCP": lhs[0]})
-    return {"$substrCP": lhs}
-
-
-def trim(operator):
-    def wrapped(self, compiler, connection):
-        lhs = process_lhs(self, compiler, connection, as_expr=True)
-        return {f"${operator}": {"input": lhs}}
-
-    return wrapped
-
-
-def trunc(self, compiler, connection):
-    lhs_mql = process_lhs(self, compiler, connection, as_expr=True)
-    lhs_mql = {"date": lhs_mql, "unit": self.kind, "startOfWeek": "mon"}
-    if timezone := self.get_tzname():
-        lhs_mql["timezone"] = timezone
-    return {"$dateTrunc": lhs_mql}
-
-
-_trunc_convert_value = TruncBase.convert_value
-
-
-def trunc_convert_value(self, value, expression, connection):
-    if connection.vendor == "mongodb":
-        # A custom TruncBase.convert_value() for MongoDB.
-        if value is None:
-            return None
-        convert_to_tz = settings.USE_TZ and self.get_tzname() != "UTC"
-        if isinstance(self.output_field, DateTimeField):
-            if convert_to_tz:
-                # Unlike other databases, MongoDB returns the value in UTC,
-                # so rather than setting the time zone equal to self.tzinfo,
-                # the value must be converted to tzinfo.
-                value = value.astimezone(self.tzinfo)
-        elif isinstance(value, datetime):
-            if isinstance(self.output_field, DateField):
-                if convert_to_tz:
-                    value = value.astimezone(self.tzinfo)
-                # Truncate for Trunc(..., output_field=DateField)
-                value = value.date()
-            elif isinstance(self.output_field, TimeField):
-                if convert_to_tz:
-                    value = value.astimezone(self.tzinfo)
-                # Truncate for Trunc(..., output_field=TimeField)
-                value = value.time()
-        return value
-    return _trunc_convert_value(self, value, expression, connection)
-
-
-def trunc_date(self, compiler, connection):
-    # Cast to date rather than truncate to date.
-    lhs_mql = process_lhs(self, compiler, connection, as_expr=True)
-    tzname = self.get_tzname()
-    if tzname and tzname != "UTC":
-        raise NotSupportedError(f"TruncDate with tzinfo ({tzname}) isn't supported on MongoDB.")
-    return {
-        "$dateFromString": {
-            "dateString": {
-                "$concat": [
-                    {"$dateToString": {"format": "%Y-%m-%d", "date": lhs_mql}},
-                    # Dates are stored with time(0, 0), so by replacing any
-                    # existing time component with that, the result of
-                    # TruncDate can be compared to DateField.
-                    "T00:00:00.000",
-                ]
-            },
-        }
-    }
-
-
-def trunc_time(self, compiler, connection):
-    tzname = self.get_tzname()
-    if tzname and tzname != "UTC":
-        raise NotSupportedError(f"TruncTime with tzinfo ({tzname}) isn't supported on MongoDB.")
-    lhs_mql = process_lhs(self, compiler, connection, as_expr=True)
-    return {
-        "$dateFromString": {
-            "dateString": {
-                "$concat": [
-                    # Times are stored with date(1, 1, 1)), so by
-                    # replacing any existing date component with that, the
-                    # result of TruncTime can be compared to TimeField.
-                    "0001-01-01T",
-                    {"$dateToString": {"format": "%H:%M:%S.%L", "date": lhs_mql}},
-                ]
-            }
-        }
-    }
-
-
-def register_functions():
-    Cast.as_mql_expr = cast
-    Concat.as_mql_expr = concat
-    ConcatPair.as_mql_expr = concat_pair
-    Cot.as_mql_expr = cot
-    Extract.as_mql_expr = extract
-    ExtractQuarter.as_mql_expr = extract_quarter
-    Func.as_mql_expr = func
-    Func.can_use_path = False
-    JSONArray.as_mql_expr = partialmethod(process_lhs, as_expr=True)
-    Left.as_mql_expr = left
-    Length.as_mql_expr = length
-    Log.as_mql_expr = log
-    Lower.as_mql_expr = preserve_null("toLower")
-    LPad.as_mql_expr = pad(is_left=True)
-    LTrim.as_mql_expr = trim("ltrim")
-    MD5.as_mql_expr = hash_func("md5")
-    Now.as_mql_expr = now
-    NullIf.as_mql_expr = null_if
-    Repeat.as_mql_expr = repeat
-    Replace.as_mql_expr = replace
-    Reverse.as_mql_expr = reverse
-    Right.as_mql_expr = right
-    Round.as_mql_expr = round_
-    RPad.as_mql_expr = pad(is_left=False)
-    RTrim.as_mql_expr = trim("rtrim")
-    SHA256.as_mql_expr = hash_func("sha256")
-    Sign.as_mql_expr = sign
-    StrIndex.as_mql_expr = str_index
-    Substr.as_mql_expr = substr
-    Trim.as_mql_expr = trim("trim")
-    TruncBase.as_mql_expr = trunc
-    TruncBase.convert_value = trunc_convert_value
-    TruncDate.as_mql_expr = trunc_date
-    TruncTime.as_mql_expr = trunc_time
-    Upper.as_mql_expr = preserve_null("toUpper")
